Split src/ directory into multiple files, plus minor config and typing changes for future React compatibility.
