+++ package.json
@@ -40,13 +40,13 @@
     "@jupyterlab/launcher": "^1.0",
     "@phosphor/coreutils": "^1.3.0",
     "@phosphor/messaging": "^1.2.2",
-    "@phosphor/widgets": "^1.6.0",
+    "@phosphor/widgets": "^1.9.0",
     "@types/bootstrap": "^4.3.0",
     "@types/datatables.net": "^1.10.12",
     "@types/datatables.net-buttons": "^1.4.0",
     "@types/datatables.net-select": "^1.2.4",
     "@types/fs-extra": "^5.0.5",
-    "@types/jquery": "^3.3.4",
+    "@types/jquery": "^3.3.31",
     "@types/node": "^11.13.7",
     "bootstrap": "^4.2.1",
     "datatables.net-buttons-dt": "^1.5.4",
@@ -54,14 +54,22 @@
     "datatables.net-plugins": "^1.10.19",
     "datatables.net-select-dt": "^1.2.4",
     "fs-extra": "^7.0.1",
-    "jquery": "^3.4",
-    "popper.js": "^1.14.7"
+    "jquery": "^3.4.1",
+    "popper.js": "^1.15.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "yarn-deduplicate": "^1.1.1"
   },
   "devDependencies": {
+    "@types/react": "^16.9.2",
+    "@types/react-dom": "^16.9.0",
     "rimraf": "^2.6.1",
-    "typescript": "~3.1.0"
+    "typescript": "~3.5.2"
   },
   "jupyterlab": {
     "extension": true
+  },
+  "resolutions": {
+  "@types/react": "^16.9.2"
   }
 }
